+++ src/snapred/ui/view/DiffCalTweakPeakView.py
@@ -21,6 +21,7 @@ from snapred.meta.Config import Config
 from snapred.meta.decorators.Resettable import Resettable
 from snapred.meta.mantid.AllowedPeakTypes import SymmetricPeakEnum
 from snapred.meta.mantid.FitPeaksOutput import FitOutputEnum
+from snapred.ui.plotting.Factory import mantidAxisFactory
 from snapred.ui.view.BackendRequestView import BackendRequestView
 from snapred.ui.widget.Toggle import Toggle
 
@@ -230,28 +231,7 @@ class DiffCalTweakPeakView(BackendRequestView):
 
             # NOTE: Mutate the ax object as the mantidaxis does not account for lines
             # TODO: Bubble this up to the mantid codebase and remove this mutation.
-            def rename_workspace(old_name, new_name):
-                """
-                Rename a workspace, and update the artists, creation arguments and tracked workspaces accordingly
-                :param new_name : the new name of workspace
-                :param old_name : the old name of workspace
-                """
-                for cargs in ax.creation_args:
-                    # NEW CHECK
-                    func_name = cargs["function"]
-                    if func_name not in ["axhline", "axvline"] and cargs["workspaces"] == old_name:
-                        cargs["workspaces"] = new_name
-                    # Alternatively,
-                    # if cargs.get("workspaces") == old_name:
-                    #     cargs["workspaces"] = new_name
-                for ws_name, ws_artist_list in list(ax.tracked_workspaces.items()):
-                    for ws_artist in ws_artist_list:
-                        if ws_artist.workspace_name == old_name:
-                            ws_artist.rename_data(new_name)
-                    if ws_name == old_name:
-                        ax.tracked_workspaces[new_name] = ax.tracked_workspaces.pop(old_name)
-
-            ax.rename_workspace = rename_workspace
+            ax = mantidAxisFactory(ax)
 
             ax.tick_params(direction="in")
             ax.set_title(f"Group ID: {wkspIndex + 1}")
